moved quiz preview
