+++ src/Kanbas/Courses/Quizzes/Preview/OpenResponse.tsx
